chore: create activeFeedNameConetxtProvider - issues with dependency flow
